Implement malware-pkg scanner

+++ internal/nexusPkgPuller/nexusPkgPuller.go
@@ -10,52 +10,52 @@ import (
 	"github.com/qiloz/nexus-shaihulud-scan/internal/utilityFlags"
 )
 
-func GetRepositoryPackages(flags utilityFlags.Flags) error {
+func GetRepositoryPackages(flags utilityFlags.Flags) (string, error) {
 	utilDirPath, err := utilEnv.CreateUtilDir()
 	if err != nil {
-		return err
+		return "", err
 	}
 
 	var pkgRepoListPath = path.Join(utilDirPath, "nx-"+flags.NxRepoName+"-npm-pkgs.csv")
 
 	f, err := os.Create(pkgRepoListPath)
 	if err != nil {
-		return err
+		return "", err
 	}
 	defer f.Close()
 
 	_, err = f.WriteString("Package Name,Version\n")
 	if err != nil {
-		return err
+		return "", err
 	}
 
 	var continuationToken string
-
+	var isFirstPage = true
 	fmt.Println("Retrieve npm packages from Nexus API...")
 	for {
 		assetData, err := nexusApiProxy.GetNxRepoAssets(flags, continuationToken)
 		if err != nil {
-			return err
+			return "", err
 		}
 
-		if len(assetData.Items) == 0 {
+		if isFirstPage && len(assetData.Items) == 0 {
 			err = f.Close()
 			if err != nil {
-				return err
+				return "", err
 			}
 
 			err = os.Remove(pkgRepoListPath)
 			if err != nil {
-				return err
+				return "", err
 			}
 
-			return fmt.Errorf("no assets found. probably repository '%s' does not exists or empty", flags.NxRepoName)
+			return "", fmt.Errorf("no assets found. probably repository '%s' does not exists or empty", flags.NxRepoName)
 		}
 
 		for _, asset := range assetData.Items {
 			_, err = f.WriteString(asset.Npm.Name + "," + asset.Npm.Version + "\n")
 			if err != nil {
-				return err
+				return "", err
 			}
 		}
 
@@ -63,10 +63,11 @@ func GetRepositoryPackages(flags utilityFlags.Flags) error {
 			break
 		}
 		fmt.Print("|")
+		isFirstPage = false
 		continuationToken = assetData.ContinuationToken
 	}
 
 	fmt.Println("\n> Repository assets was written to: " + pkgRepoListPath)
 
-	return nil
+	return pkgRepoListPath, nil
 }
